Polish OpenDreamer lifecycle and schema
A playable OpenDreamer world is scoped to a Reactor session, so initialize and release its shared state with session lifecycle hooks instead of resetting it for every viewer connection. Keep checkpoint and GPU setup in load(), while releasing uploaded conditioning when the session ends.

Separate public schema types and integration utilities from the inference pipeline, clarify the generated command and message descriptions, and pin the recipe to reactor-runtime 3.1.2.
