+++ examples/open-dreamer/opendreamer_types.py
@@ -1,0 +1,120 @@
+"""Define OpenDreamer's configuration and public Reactor schema."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+from reactor_runtime import (
+    InputField,
+    InputState,
+    MessageField,
+    ModelMessage,
+    Output,
+    Video,
+)
+
+DEMO_CHOICES = ["demo_1", "demo_2", "demo_3"]
+
+
+@dataclass(frozen=True)
+class DemoConfig:
+    """Describe a dataset window available as a starting scene."""
+
+    name: str
+    video: Path
+    actions: Path
+    start_frame: int
+
+
+@dataclass(frozen=True)
+class OpenDreamerConfig:
+    """Hold validated model, checkpoint, and conditioning settings."""
+
+    source_revision: str
+    checkpoint_repo_id: str
+    checkpoint_revision: str
+    platform: str
+    seed: int
+    num_steps: int
+    tau_ctx_target: float
+    conditioning_frames: int
+    demos: tuple[DemoConfig, ...]
+    warmup_steps: int
+    memory_fraction: float
+
+
+@dataclass(frozen=True)
+class RolloutConditioning:
+    """Pair consecutive Minecraft frames with their aligned player actions."""
+
+    frames: np.ndarray
+    actions: Any
+
+
+class OpenDreamerOutput(Output):
+    """Stream the next Minecraft frame generated by OpenDreamer."""
+
+    main_video: Video
+
+
+class ActionChanged(ModelMessage):
+    """Report the player controls OpenDreamer will apply to generated frames."""
+
+    control: str = MessageField(
+        description="Name of the command that updated the controls."
+    )
+    paused: bool = MessageField(description="Whether generation is currently paused.")
+    pressed_keys: list[str] = MessageField(
+        description="Keyboard keys held for each generated frame until released."
+    )
+    pressed_mouse_buttons: list[str] = MessageField(
+        description="Mouse buttons held for each generated frame until released."
+    )
+    delta_x: float = MessageField(
+        description="Horizontal mouse movement queued for the next generated frame."
+    )
+    delta_y: float = MessageField(
+        description="Vertical mouse movement queued for the next generated frame."
+    )
+    wheel_delta: int = MessageField(
+        description="Mouse-wheel movement queued for the next generated frame."
+    )
+
+
+class ConditioningChanged(ModelMessage):
+    """Report the starting scene selected for the next rollout."""
+
+    source: str = MessageField(
+        description='Starting-scene source: "demo" for a dataset sample or "upload" for an image.'
+    )
+    selection: str = MessageField(
+        description="Selected demo name or uploaded image filename."
+    )
+
+
+class RolloutReset(ModelMessage):
+    """Report the seed and starting scene selected for a rollout reset."""
+
+    seed: int = MessageField(description="Random seed the restarted rollout will use.")
+    conditioning: str = MessageField(
+        description='Selected demo name, or "uploaded" when using an uploaded image.'
+    )
+
+
+class OpenDreamerState(InputState):
+    """Expose the controls shared by one playable OpenDreamer world."""
+
+    paused: bool = InputField(
+        default=False,
+        description="Pause frame generation while preserving the current world state.",
+    )
+    _pressed_keys: frozenset[str] = frozenset()
+    _pressed_mouse_buttons: frozenset[str] = frozenset()
+    _delta_x: float = 0.0
+    _delta_y: float = 0.0
+    _wheel_delta: int = 0
+    _reset_requested: bool = True
+    _seed: int = 0
